Assignment/Assignment08/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
cssselect==1.2.0
fake-useragent==2.0.3
idna==3.10
importlib_metadata==8.6.1
lxml==5.3.1
lxml_html_clean==0.4.1
meta_ai_api==1.2.5
parse==1.20.2
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
requests==2.32.3
requests-html==0.10.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
uv==0.1.25
w3lib==2.3.1
websockets==10.4
zipp==3.21.0
PyPDF2
python-dotenv==0.19.0
langchain-google-genai
google-generativeai
langchain
